mudança de variaveis de dados

+++ src/aula06/exercicios/Exercicio4.java
@@ -4,12 +4,12 @@ public class Exercicio4 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        double[] notas = new double[4];
-        double soma = 0;
+        int[] notas = new int[4];
+        int soma = 0;
 
         for (int i = 0; i < notas.length; i++) {
             System.out.print("Digite a nota " + (i + 1) + ": ");
-            notas[i] = scanner.nextDouble(); 
+            notas[i] = scanner.nextInt(); 
             soma += notas[i]; 
        }
         double media = soma / notas.length;
